update syncthing and the UI
this is quite a big change.
The sources have been generated from scratch with custom generators.
Both the python and the go generator needed changes to accommodate
syncthing.

The Go sources can be generated by sharing the network and calling go
build -mod=vendor. The resulting vendor/modules.txt needs to be run
through the custom go generator.

The patches have been included upstream. But a new one was written for
the appdata file.